requirements.txt
flask
gunicorn
python-binance
jsonify